Moved non-gui files to libs module
