+++ libs/creatures.py
