+++ qtheader.py
@@ -1,7 +1,7 @@
 from PySide.QtGui import QApplication, QWidget, QGridLayout, QPushButton, QPixmap, QLabel, QPalette, QSizePolicy, QTextEdit, QComboBox, QCheckBox, QLineEdit, QListWidget, QLayout, QIcon, QMessageBox, QDesktopWidget
 from PySide.QtCore import Qt, QRect
 from popup import popup
-import json_reader
+import libs.json_reader as json_reader
 
 # This software is provided 'as-is'. Objeus and it's members are not responsible for any damages
 # caused by the use of this software.
