+++ qtslgui.py
@@ -1,6 +1,6 @@
 from qtheader import *
-from sls import SocialLink
-from action import *
+from libs.sls import SocialLink
+from libs.action import *
 from simulate import Simulation
 from slview import PrettySL
 from slinfo import LinkInfo
